requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.5
djangorestframework==3.14.0
drf-spectacular==0.26.5
Faker==19.6.2
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
mysqlclient==2.1.1
packaging==23.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.6
